Update to command line scripts for Android 9, 10, 11 without root.
Use create-debuggable-apk.sh script to take treehearder apk's for
fenix-nightly and turn them into debuggable apks.

Use create-debuggable-apk-wrapper.sh to automate this with chron.

Use reinstall-fenix-as-debug-apk.sh to automate replacement of
fenix-nightly binaries on attached devices to debug versions of the same.

Use fenix-nightly.sh to run TOT browsertime with TOT geckodriver, using
--android-storage=app.

Use (run-once-per-day, run-through-sites, strip-url-to-tld) to automate
running of browsertime on devices with all the above in place.